Updated to blazor server and added mudblazor
