+++ BlazorQueryBuilder/Components/Layout/NavMenu.razor
@@ -1,0 +1,3 @@
+﻿<MudNavMenu>
+    <MudNavLink Href="" Match="NavLinkMatch.All" Icon="@Icons.Material.Filled.Build">Query Builder</MudNavLink>
+</MudNavMenu>
