+++ BlazorQueryBuilder/Pages/QueryBuilder.razor
@@ -7,9 +7,10 @@
 @using BlazoryQueryBuilder.Shared.Models
 @using BlazoryQueryBuilder.Shared.Services
 @using System.Reflection
+@using System.Collections
 @inject PredicateFactory PredicateFactory
-@inject HttpClient HttpClient
 @inject QueryBuilderService<TEntity> QueryBuilderService
+@inject QueryServiceFactory<MyDbContext> QueryServiceFactory
 @typeparam TEntity
 
 <div class="form-group">
@@ -25,8 +26,8 @@
 <div class="row">
     <div class="col-12">
         <LambdaComponent OnChanged="@OnChanged"
-                         Lambda="@QueryBuilderService.Lambda"
-                         Parameter="@QueryBuilderService.Parameter">
+        Lambda="@QueryBuilderService.Lambda"
+        Parameter="@QueryBuilderService.Parameter">
         </LambdaComponent>
     </div>
 </div>
@@ -52,9 +53,6 @@
 </div>
 
 @code {
-
-
-
     List<TEntity> _data = new List<TEntity>();
 
     Expression<Func<TEntity, bool>> Predicate => (Expression<Func<TEntity, bool>>)QueryBuilderService.Lambda;
@@ -62,6 +60,11 @@
 
     [Parameter]
     public string Expression { get; set; }
+
+    protected override Task OnInitializedAsync()
+    {
+        return base.OnInitializedAsync();
+    }
 
     protected override void OnInitialized()
     {
@@ -92,15 +95,15 @@
     private async Task RunQuery()
     {
         Func<TEntity, bool> predicate = Predicate.Compile();
-        var content = JsonContent.Create(new BlazoryQueryBuilder.Shared.Models.Predicate
-            {
+        var predicateModel = new BlazoryQueryBuilder.Shared.Models.Predicate
+        {
                 LambdaExpression = Predicate.ToString(),
                 EntityType = typeof(TEntity).Name,
                 SelectedProperties = SelectedProperties.Select(prop => prop.Name).ToList()
-            });
-        var data = await HttpClient.PostAsync("https://localhost:44338/api/values", content);
+        };
+        var data = await ExecuteQuery(predicateModel) as IEnumerable<TEntity>;
 
-        _data = (await data.Content.ReadFromJsonAsync<IEnumerable<TEntity>>()).ToList();
+        _data = data.ToList();
     }
 
     private void OnChanged(Expression body)
@@ -110,4 +113,10 @@
         StateHasChanged();
     }
 
+    public async Task<IEnumerable> ExecuteQuery(BlazoryQueryBuilder.Shared.Models.Predicate predicate)
+    {
+        IQueryService queryService = QueryServiceFactory.Create(predicate.EntityType);
+        IEnumerable data = await queryService.QueryData(predicate.LambdaExpression, predicate.SelectedProperties);
+        return data;
+    }
 }
