+++ BlazorQueryBuilder/Pages/QueryBuilderMain.razor
@@ -1,30 +1,30 @@
 ﻿@page "/query-builder-main"
+@page "/"
 @using System.Reflection
 @using BlazoryQueryBuilder.Shared.Models
 @using System.Linq.Dynamic.Core
 @using System.Linq.Expressions
+@using MudBlazor
 
 <div class="row">
     <div class="col-12">
         @if (_queryLoaded || _newQuery)
         {
-            <button class="btn btn-primary" @onclick="@StartOver">Start Over</button>
+            <MudButton Color="Color.Primary" OnClick="@StartOver">Start Over</MudButton>
         }
         else
         {
-            <button class="btn btn-primary" @onclick="@LoadQuery">Load Query</button>
-            <button class="btn btn-primary" @onclick="@NewQuery">New Query</button>
+            <MudButton Color="Color.Primary" OnClick="@LoadQuery">Load Query</MudButton>
+            <MudButton Color="Color.Primary" OnClick="NewQuery">New Query</MudButton>
         }
     </div>
 </div>
 
-
-
 @if (_newQuery)
 {
-    <div class="row">
-        <div class="col-6">
-            <h1>Entity</h1>
+    <MudStack Spacing="2">
+        <MudItem xs="12">
+            <MudText Typo="Typo.h4">Entity</MudText>
             <div class="form-group">
                 <select class="form-control" id="exampleFormControlSelect1" @onchange="@Select">
                     @foreach (var entity in Entities)
@@ -33,14 +33,11 @@
                     }
                 </select>
             </div>
-        </div>
-    </div>
+        </MudItem>
+    </MudStack>
 }
 
-
 @CreateQueryBuilder()
-
-
 
 @code {
