--- BlazorQueryBuilder/Pages/_Imports.razor
@@ -1,1 +1,0 @@
-﻿@layout MainLayout
