+++ BlazorQueryBuilder/Routes.razor
@@ -1,0 +1,7 @@
+﻿@using BlazorQueryBuilder.Components.Layout
+<Router AppAssembly="@typeof(App).Assembly">
+    <Found Context="routeData">
+        <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
+        <FocusOnNavigate RouteData="@routeData" Selector="h1" />
+    </Found>
+</Router>
