--- BlazorQueryBuilder/Shared/SurveyPrompt.razor
@@ -1,15 +1,0 @@
-﻿<div class="alert alert-secondary mt-4" role="alert">
-    <span class="oi oi-pencil mr-2" aria-hidden="true"></span>
-    <strong>@Title</strong>
-
-    <span class="text-nowrap">
-        Please take our
-        <a target="_blank" class="font-weight-bold" href="https://go.microsoft.com/fwlink/?linkid=2026993">brief survey</a>
-    </span>
-    and tell us what you think.
-</div>
-
-@code {
-    [Parameter]
-    public string Title { get; set; } // Demonstrates how a parent component can supply parameters
-}
