+++ BlazorQueryBuilder/_Imports.razor
@@ -3,6 +3,9 @@
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
+@using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using BlazorQueryBuilder.Shared
+@using MudBlazor
+@using BlazorQueryBuilder
+@using BlazoryQueryBuilder
