+++ BlazorQueryBuilder/appsettings.json
@@ -1,0 +1,3 @@
+{
+  "DetailedErrors": true
+}
